implement authorization id for SASL
Expose the SASL authorization id request via a cmdline flag.

Using `-authz-id foo@REALM` or `-authz-id foo` works with Cyrus IMAP in
this setup, but `-authz-id foo@realm` (lowercase) does not, not does
another value of foo not allowed by server config.  Absent (empty)
authz-id continues to work, taking the default.

Bad authz-id values result in the server logging:

    authentication failure: user foo is not allowed to proxy

and a client panic, which is part of the general poor error recovery of
this tool so far.