Apcera's GSSAPI supports ContinueNeeded now
Instead of doing step-counting which only works for Kerberos, with hard
knowledge of step-counts needed, we can now use the ContinueNeeded
"error" return from the `gssapi` import.

(If we needed to, we could also reference `a.lib.LastStatus` to get the
 real status response of the last action, even if it's not an error)

This library pseudo-error forces us, as the client, to understand the
continue-needed concept and not shoot ourselves in the feet by only
doing one pass through the auth because it's not considered.

Cunning Russian programmer FTW.